+++ x/msgfees/client/cli/cli_test.go
@@ -71,75 +71,134 @@ func (s *IntegrationTestSuite) TearDownSuite() {
 
 func (s *IntegrationTestSuite) TestMsgFeesTxGovProposals() {
 	testCases := []struct {
-		name         string
-		args         []string
-		expectErr    bool
-		expectErrMsg string
-		expectedCode uint32
+		name          string
+		typeArg       string
+		title         string
+		description   string
+		deposit       string
+		msgType       string
+		additionalFee string
+		recipient     string
+		bips          string
+		expectErrMsg  string
+		expectedCode  uint32
 	}{
-		{"add msg based fee proposal - valid",
-			[]string{
-				"add",
-				"test add msg based fee",
-				"description",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-				"--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
-				fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
-			},
-			false,
-			"",
-			0,
+		{
+			name:          "add msg based fee proposal - valid",
+			typeArg:       "add",
+			title:         "test add msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
 		},
-		{"add msg based fee proposal - invalid msg type url",
-			[]string{
-				"add",
-				"test add msg based fee",
-				"description",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-				"--msg-type=invalid",
-				fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
-			},
-			true,
-			"unable to resolve type URL invalid",
-			0,
+		{
+			name:          "add msg based fee proposal with recipient - valid",
+			typeArg:       "add",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
 		},
-		{"add msg based fee proposal - invalid additional fee",
-			[]string{
-				"add",
-				"test add msg based fee",
-				"description",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-				"--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
-				fmt.Sprintf("--additional-fee=%s", "blah"),
-			},
-			true,
-			"invalid decimal coin expression: blah",
-			0,
+		{
+			name:          "add msg based fee proposal with recipient and default basis points - valid",
+			typeArg:       "add",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     fmt.Sprintf("--recipient=%s", s.testnet.Validators[0].Address.String()),
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
 		},
-		{"update msg based fee proposal - valid",
-			[]string{
-				"update",
-				"test update msg based fee",
-				"description",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-				"--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
-				fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
-			},
-			false,
-			"",
-			0,
+		{
+			name:          "add msg based fee proposal - invalid msg type url",
+			typeArg:       "add",
+			title:         "test add msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=invalid",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "unable to resolve type URL invalid",
+			expectedCode:  0,
 		},
-		{"remove msg based fee proposal - valid",
-			[]string{
-				"remove",
-				"test update msg based fee",
-				"description",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-				"--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
-			},
-			false,
-			"",
-			0,
+		{
+			name:          "add msg based fee proposal - invalid additional fee",
+			typeArg:       "add",
+			title:         "test add msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", "blah"),
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "invalid decimal coin expression: blah",
+			expectedCode:  0,
+		},
+		{
+			name:          "update msg based fee proposal with recipient - valid",
+			typeArg:       "update",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
+		},
+		{
+			name:          "update msg based fee proposal with recipient and default basis points - valid",
+			typeArg:       "update",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     fmt.Sprintf("--recipient=%s", s.testnet.Validators[0].Address.String()),
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
+		},
+		{
+			name:          "update msg based fee proposal with recipient and modified basis points - valid",
+			typeArg:       "update",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: fmt.Sprintf("--additional-fee=%s", sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(12))).String()),
+			recipient:     fmt.Sprintf("--recipient=%s", s.testnet.Validators[0].Address.String()),
+			bips:          fmt.Sprintf("--bips=%s", "5001"),
+			expectErrMsg:  "",
+			expectedCode:  0,
+		},
+		{
+			name:          "remove msg based fee proposal - valid",
+			typeArg:       "remove",
+			title:         "test update msg based fee",
+			description:   "description",
+			deposit:       sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			msgType:       "--msg-type=/provenance.metadata.v1.MsgWriteRecordRequest",
+			additionalFee: "",
+			recipient:     "",
+			bips:          "",
+			expectErrMsg:  "",
+			expectedCode:  0,
 		},
 	}
 
@@ -155,10 +214,20 @@ func (s *IntegrationTestSuite) TestMsgFeesTxGovProposals() {
 				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
 				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
 			}
-			tc.args = append(tc.args, args...)
 
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, msgfeescli.GetCmdMsgFeesProposal(), tc.args)
-			if tc.expectErr {
+			args = append(args, tc.typeArg, tc.title, tc.description, tc.deposit, tc.msgType)
+			if len(tc.additionalFee) != 0 {
+				args = append(args, tc.additionalFee)
+			}
+			if len(tc.recipient) != 0 {
+				args = append(args, tc.recipient)
+			}
+			if len(tc.bips) != 0 {
+				args = append(args, tc.bips)
+			}
+
+			out, err := clitestutil.ExecTestCLICmd(clientCtx, msgfeescli.GetCmdMsgFeesProposal(), args)
+			if len(tc.expectErrMsg) != 0 {
 				s.Require().Error(err)
 				s.Assert().Equal(tc.expectErrMsg, err.Error())
 			} else {
@@ -172,43 +241,39 @@ func (s *IntegrationTestSuite) TestMsgFeesTxGovProposals() {
 func (s *IntegrationTestSuite) TestUpdateUsdConversionRateProposal() {
 	testCases := []struct {
 		name         string
-		args         []string
-		expectErr    bool
+		title        string
+		description  string
+		rate         string
+		deposit      string
 		expectErrMsg string
 		expectedCode uint32
 	}{
-		{"update nhash to usd mil proposal - valid",
-			[]string{
-				"title",
-				"description",
-				"10",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-			},
-			false,
-			"",
-			0,
+		{
+			name:         "update nhash to usd mil proposal - valid",
+			title:        "title",
+			description:  "description",
+			rate:         "10",
+			deposit:      sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			expectErrMsg: "",
+			expectedCode: 0,
 		},
-		{"update nhash to usd mil proposal - invalid - rate param error",
-			[]string{
-				"title",
-				"description",
-				"invalid-rate",
-				sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
-			},
-			true,
-			"unable to parse nhash value: invalid-rate",
-			0,
+		{
+			name:         "update nhash to usd mil proposal - invalid - rate param error",
+			title:        "title",
+			description:  "description",
+			rate:         "invalid-rate",
+			deposit:      sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String(),
+			expectErrMsg: "unable to parse nhash value: invalid-rate",
+			expectedCode: 0,
 		},
-		{"update nhash to usd mil proposal - invalid - deposit param",
-			[]string{
-				"title",
-				"description",
-				"10",
-				"invalid-deposit",
-			},
-			true,
-			"invalid decimal coin expression: invalid-deposit",
-			0,
+		{
+			name:         "update nhash to usd mil proposal - invalid - deposit param",
+			title:        "title",
+			description:  "description",
+			rate:         "10",
+			deposit:      "invalid-deposit",
+			expectErrMsg: "invalid decimal coin expression: invalid-deposit",
+			expectedCode: 0,
 		},
 	}
 
@@ -223,10 +288,10 @@ func (s *IntegrationTestSuite) TestUpdateUsdConversionRateProposal() {
 				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
 				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
 			}
-			tc.args = append(tc.args, args...)
+			args = append(args, tc.name, tc.description, tc.rate, tc.deposit)
 
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, msgfeescli.GetUpdateNhashPerUsdMilProposal(), tc.args)
-			if tc.expectErr {
+			out, err := clitestutil.ExecTestCLICmd(clientCtx, msgfeescli.GetUpdateNhashPerUsdMilProposal(), args)
+			if len(tc.expectErrMsg) != 0 {
 				s.Require().Error(err)
 				s.Assert().Equal(tc.expectErrMsg, err.Error())
 			} else {